Fix a few more CSS issues; top header = link to /
In portrait mode, remove the side margins on articles, to allow more
text to fit the width of the page. Fix the ridiculous amount of left
indentation on in-article ordered & unordered lists, and make the
margins match normal paragraphs. Correctly hide the in-page hamburger
menu in portrait mode. Make top header a link back to /, and make it
slightly smaller in landscape mode.